# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST complète pour la gestion d'une plateforme éducative avec familles, enseignants et administrateurs",
        "title": "API REST Go - Plateforme Éducative",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/options": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère la liste des options avec possibilité de filtrage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Liste toutes les options",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Statut de l'option",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID de l'enseignant",
                        "name": "enseignant_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID de la famille",
                        "name": "famille_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID de l'offre",
                        "name": "offer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Option"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crée une nouvelle option sur une offre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Création d'une option",
                "parameters": [
                    {
                        "description": "Données de l'option",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OptionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/options/expiring": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère la liste des options qui vont bientôt expirer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Liste des options expirant bientôt",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Option"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/options/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère la liste des options avec le statut \"en attente\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Liste des options en attente",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Option"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/options/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Met à jour les informations d'une option existante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Mise à jour d'une option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'option",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Données de mise à jour",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OptionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Supprime une option existante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Suppression d'une option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'option",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/options/{id}/cancel": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Annule une option acceptée",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Annulation d'une option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'option",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/options/{id}/decline": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refuse une option en attente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Refus d'une option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'option",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/options/{id}/expire": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marque une option comme expirée",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Expiration d'une option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'option",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/options/{id}/reject": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Rejette une option en attente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Rejet d'une option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'option",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.OptionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enseignant": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Enseignant"
                        }
                    ]
                },
                "enseignant_id": {
                    "description": "Foreign Keys",
                    "type": "integer"
                },
                "expiration_date": {
                    "type": "string"
                },
                "famille": {
                    "$ref": "#/definitions/models.Famille"
                },
                "famille_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "offer": {
                    "$ref": "#/definitions/models.Offer"
                },
                "offer_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.OptionStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Enseignant": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Course"
                    }
                },
                "missions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Mission"
                    }
                },
                "offers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Offer"
                    }
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Option"
                    }
                },
                "qualifications": {
                    "type": "string"
                },
                "reports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Report"
                    }
                },
                "specialization": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Option": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enseignant": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Enseignant"
                        }
                    ]
                },
                "enseignant_id": {
                    "description": "Foreign Keys",
                    "type": "integer"
                },
                "expiration_date": {
                    "type": "string"
                },
                "famille": {
                    "$ref": "#/definitions/models.Famille"
                },
                "famille_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "offer": {
                    "$ref": "#/definitions/models.Offer"
                },
                "offer_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.OptionStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.OptionCreateRequest": {
            "type": "object",
            "required": [
                "enseignant_id",
                "famille_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "enseignant_id": {
                    "type": "integer"
                },
                "expiration_date": {
                    "type": "string"
                },
                "famille_id": {
                    "type": "integer"
                },
                "offer_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.OptionUpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "expiration_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.OptionStatus"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}